+++ src/geocoord/geo.rs
@@ -15,11 +15,13 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 extern crate regex;
+extern crate chrono;
 
 use std::f64::consts;
 use std::fmt;
-//use std::cmp::*;
-use self::regex::Regex;
+use std::time;
+//use std::iter;
+use self::regex::{Regex};
 
 // ---- PixelPos -----------------------------------------------------------------------------------
 
@@ -45,24 +47,43 @@ impl fmt::Display for PixelPos {
 
 // ---- Location -----------------------------------------------------------------------------------
 
-/// Map coordinates in degrees.
+/// A struct representing geographic location and point of time. 
+/// The coordinates are expected to use WGS84 reference ellipsoid.
+/// The altitude is in metres above sea leve.
 #[derive(Copy, Clone)]
 pub struct Location {
     pub lat: f64, // south-north
     pub lon: f64, // west-east
+    pub elevation: Option<f64>, // altitude in metres
+    pub time: Option<chrono::DateTime<chrono::UTC>>, // time related to the location
 }
 
 impl Location {
-    // Constructor
+    // Constructor with coordinates only.
     pub fn new(lat: f64, lon: f64) -> Location {
-        Location{lat: lat, lon: lon}
-    }
-
-    pub fn from_str(lat_lon_str: &str) -> Result<Location, String> {
-        Location::from_string(lat_lon_str.to_string())
+        Location{lat: lat, lon: lon, elevation: None, time: None}
     }
     
-    pub fn from_string(lat_lon_str: String) -> Result<Location, String> {
+    // Constructor with coordinates only.
+    pub fn new_with_elevation(lat: f64, lon: f64, elevation: f64) -> Location {
+        Location{lat: lat, lon: lon, elevation: Some(elevation), time: None}
+    }
+    
+    // Constructor with coordinates only.
+    pub fn new_with_elevation_and_time(lat: f64, lon: f64, elevation: f64, time: chrono::DateTime<chrono::UTC>) -> Location {
+        Location{lat: lat, lon: lon, elevation: Some(elevation), time: Some(time)}
+    }
+    
+    // Constructor with all fields.
+    pub fn new_with_all(lat: f64, lon: f64, elevation: Option<f64>, time: Option<chrono::DateTime<chrono::UTC>>) -> Location {
+        Location{lat: lat, lon: lon, elevation: elevation, time: time}
+    }
+
+    pub fn new_with_str(lat_lon_str: &str) -> Result<Location, String> {
+        Location::new_with_string(lat_lon_str.to_string())
+    }
+    
+    pub fn new_with_string(lat_lon_str: String) -> Result<Location, String> {
         let fre = "[0-9]*\\.[0-9]+|[0-9]+";
         //let sfre = "-?[0-9]*\\.[0-9]+|-?[0-9]+";
         
@@ -77,14 +98,14 @@ impl Location {
             let ns = { if caps.name("latside").expect("unexpected") == "N" { 1.0 } else { -1.0 }  };
             let ew = { if caps.name("lonside").expect("unexpected") == "E" { 1.0 } else { -1.0 }  };
             if lat.is_some() && lon.is_some() {
-                return Ok(Location{lat: ns * lat.expect("unexpected").parse::<f64>().unwrap(), lon: ew * lon.expect("unexpected").parse::<f64>().unwrap()});
+                return Ok(Location::new(ns * lat.expect("unexpected").parse::<f64>().unwrap(), ew * lon.expect("unexpected").parse::<f64>().unwrap()));
             } 
         }
         Err(format!("bad location: {}", lat_lon_str))
     }
 
     /// True if this location is east from the other location.
-    pub fn east_from(&self, other: Location) -> bool {
+    pub fn east_from(&self, other: &Location) -> bool {
         let lon = self.lon;
         let olon = other.lon;
         
@@ -100,22 +121,22 @@ impl Location {
     }
     
     /// True if this location is west from the other location.
-    pub fn west_from(&self, other: Location) -> bool {
+    pub fn west_from(&self, other: &Location) -> bool {
         ! self.east_from(other)
     }
 
     /// True if this location is north from the other location.
-    pub fn north_from(&self, other: Location) -> bool {
+    pub fn north_from(&self, other: &Location) -> bool {
         self.lat > other.lat
     }
 
     /// True if this location is south from the other location.
-    pub fn south_from(&self, other: Location) -> bool {
+    pub fn south_from(&self, other: &Location) -> bool {
         self.lat < other.lat
     }
     
-    /// Distance to the other location.
-    pub fn distance_to(&self, other: Location) -> f64 {
+    /// Distance to the other location in metres (m).
+    pub fn distance_to(&self, other: &Location) -> f64 {
         // See: http://www.movable-type.co.uk/scripts/latlong.html
         const R: f64 = 6371000.0;
         let d_lat = other.lat - self.lat;
@@ -129,7 +150,7 @@ impl Location {
     }
     
     /// Bearing to the given location.
-    pub fn bearing_to(&self, other: Location) -> f64 {
+    pub fn bearing_to(&self, other: &Location) -> f64 {
         // See: http://www.movable-type.co.uk/scripts/latlong.html
         let d_lon = other.lon - self.lon;
         let y = deg_sin(d_lon) * deg_cos(other.lat);
@@ -148,6 +169,28 @@ impl Location {
                    cos(dr) - deg_sin(self.lat) * deg_sin(lat2));
         Location::new(lat2, lon2)
     }   
+    
+    /// Returns the average speed between the points in metres per second (m/s).
+    /// None is returned if time is missing from either of the points.
+    pub fn average_speed(&self, other: &Location) -> Option<f64> {
+        let s = self.distance_to(other);
+        if let Some(t) = self.delta_time(other) {
+            return Some(s / t);
+        }
+        None
+    }
+    
+    /// Returns delta time in seconds. If either of the points doesn't have time defined
+    /// return None.
+    pub fn delta_time(&self, other: &Location) -> Option<f64> {
+        if let Some(time0) = self.time {
+            if let Some(time1) = other.time {
+                let duration = (time1 - time0).to_std().unwrap_or(time::Duration::from_secs(0));
+                return Some(duration.as_secs() as f64 + 0.000000001 * duration.subsec_nanos() as f64);
+            }
+        }
+        None
+    }
 }
 
 impl fmt::Display for Location {
@@ -167,6 +210,17 @@ impl fmt::Display for Location {
         }
     }
 }
+
+// ---- LocationSequence ---------------------------------------------------------------------------
+
+// Abstraction of location sequence, also known as multi line, route, track, path, etc.
+pub trait LocationSequence {
+    fn bounding_box(&self) -> GeoBox;
+
+    fn iterator<L: Iterator>(&self) -> L;
+}
+
+
 
 // ---- GeoBox -------------------------------------------------------------------------------------
 
@@ -204,45 +258,45 @@ impl GeoBox {
     }
     
     /// True if the given location is completely inside this box.
-    pub fn contains(&self, loc: Location) -> bool {
-        if self.northwest.west_from(self.southeast) {
+    pub fn contains(&self, loc: &Location) -> bool {
+        if self.northwest.west_from(&self.southeast) {
             // Normal "small" areas
-            loc.south_from(self.northwest) && loc.north_from(self.southeast) &&
-            loc.west_from(self.southeast)  && loc.east_from(self.northwest)
+            loc.south_from(&self.northwest) && loc.north_from(&self.southeast) &&
+            loc.west_from(&self.southeast)  && loc.east_from(&self.northwest)
         } else {
             // Bigger than semi-equator areas
-            loc.south_from(self.northwest) && loc.north_from(self.southeast) &&
-            !(loc.east_from(self.southeast) && loc.west_from(self.northwest))
+            loc.south_from(&self.northwest) && loc.north_from(&self.southeast) &&
+            !(loc.east_from(&self.southeast) && loc.west_from(&self.northwest))
         }
     }
     
     /// True if the given and this box have common area.
-    pub fn intersects(&self, other: GeoBox) -> bool {
-        self.contains(other.northwest) ||
-        self.contains(other.southeast) ||
-        other.contains(self.northwest) ||
-        other.contains(self.southeast) ||
-        self.contains(other.northeast()) ||
-        self.contains(other.southwest()) ||
-        other.contains(self.northeast()) ||
-        other.contains(self.southwest())
+    pub fn intersects(&self, other: &GeoBox) -> bool {
+        self.contains(&other.northwest) ||
+        self.contains(&other.southeast) ||
+        other.contains(&self.northwest) ||
+        other.contains(&self.southeast) ||
+        self.contains(&other.northeast()) ||
+        self.contains(&other.southwest()) ||
+        other.contains(&self.northeast()) ||
+        other.contains(&self.southwest())
     }
     
     /// Expand the area, to cover the given location too.
-    pub fn expand(&self, loc: Location) -> GeoBox {
+    pub fn expand(&self, loc: &Location) -> GeoBox {
         let mut gb = *self;
 
 	    // Longitude
-	    if loc.west_from(self.northwest) {
+	    if loc.west_from(&self.northwest) {
 	        gb.northwest = Location::new(gb.northwest.lat, loc.lon);
-	    } else if loc.east_from(self.southeast) {
+	    } else if loc.east_from(&self.southeast) {
 	        gb.southeast = Location::new(gb.southeast.lat, loc.lon);
 	    }
 	    
 	    // Latitude
-	    if loc.north_from(self.northwest) {
+	    if loc.north_from(&self.northwest) {
 	        gb.northwest = Location::new(loc.lat, gb.northwest.lon);
-	    } else if loc.south_from(self.southeast) {
+	    } else if loc.south_from(&self.southeast) {
 	        gb.southeast = Location::new(loc.lat, gb.southeast.lon);
 	    }
 	    
@@ -255,43 +309,6 @@ impl fmt::Display for GeoBox {
         write!(f, "{} - {}", self.northwest.to_string(), self.southeast.to_string())
     }
 }
-
-/*
-impl Ord for GeoBox {
-    /// Ordering based on intersecting. If the boxes intersect they are considered to be equal.
-    fn cmp(&self, other: &Self) -> Ordering {
-        if self.southeast.lat < other.northwest.lat {
-            Ordering::Less
-        } else if self.northwest.lat > other.southeast.lat {
-            Ordering::Greater
-        } else {
-            if self.southeast.lon < other.northwest.lon {
-                Ordering::Less
-            } else if self.northwest.lon > other.southeast.lon {
-                Ordering::Greater
-            } else {
-                Ordering::Equal
-            }
-        }
-    }
-}
-
-impl PartialOrd for GeoBox {
-    /// Ordering based on intersecting.
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl PartialEq for GeoBox {
-    /// Ordering based on intersecting.
-    fn eq(&self, other: &Self) -> bool {
-        self.cmp(other) == Ordering::Equal
-    }
-}
-
-impl Eq for GeoBox {}
-*/
 
 // ---- Projection ---------------------------------------------------------------------------------
 
@@ -399,7 +416,7 @@ mod tests {
     fn close_enough_to(val: f64, expected: f64, max_error: f64) -> bool { (val - expected).abs() <= max_error }
 
     #[test]
-    fn test_locations() {
+    fn test_location_directions() {
         // Locations
         let paris             = Location::new(48.8567, 2.3508);
         let visby             = Location::new(57.634722, 18.299167);
@@ -408,120 +425,134 @@ mod tests {
         let ushuaia           = Location::new(-54.8, -68.3);
         let willemstad        = Location::new(12.116667, -68.933333);
         let unalaska          = Location::new(53.880857, 166.542134);
-        let wrangel_west      = Location::from_str("71°N 179°E").unwrap();
-        let wrangel_east      = Location::from_str("71°N 178°W").unwrap();
+        let wrangel_west      = Location::new_with_str("71°N 179°E").unwrap();
+        let wrangel_east      = Location::new_with_str("71°N 178°W").unwrap();
 
         // Direction tests
-        assert!( visby.east_from(paris) );
-        assert!( paris.south_from(visby) );
-        assert!( visby.north_from(paris) );
-        assert!( melbourne.west_from(willemstad) );
-        assert!( melbourne.west_from(unalaska) );
-        assert!( unalaska.west_from(willemstad) );
-        assert!( willemstad.west_from(paris) );
-        assert!( paris.west_from(port_aux_francais) );
-        assert!( port_aux_francais.west_from(melbourne) );
-        assert!( melbourne.south_from(visby) );
+        assert!( visby.east_from(&paris) );
+        assert!( paris.south_from(&visby) );
+        assert!( visby.north_from(&paris) );
+        assert!( melbourne.west_from(&willemstad) );
+        assert!( melbourne.west_from(&unalaska) );
+        assert!( unalaska.west_from(&willemstad) );
+        assert!( willemstad.west_from(&paris) );
+        assert!( paris.west_from(&port_aux_francais) );
+        assert!( port_aux_francais.west_from(&melbourne) );
+        assert!( melbourne.south_from(&visby) );
         
         assert!( ushuaia.to_string() == "54.8°S 68.3°W" );
         
-        let ushuaia2 = Location::from_string(ushuaia.to_string()).unwrap();
+        let ushuaia2 = Location::new_with_string(ushuaia.to_string()).unwrap();
         assert!( ushuaia2.to_string() == "54.8°S 68.3°W" );
         
-        assert!( Location::from_string("nonsense".to_string()).is_err() );
+        assert!( Location::new_with_string("nonsense".to_string()).is_err() );
         
         // Distance, bearing and direction tests
-        assert!( wrangel_west.west_from(wrangel_east) );
-        assert!( wrangel_east.east_from(wrangel_west) );
+        assert!( wrangel_west.west_from(&wrangel_east) );
+        assert!( wrangel_east.east_from(&wrangel_west) );
 
-        assert!( close_enough_to(wrangel_west.distance_to(wrangel_east), 108000.0, 1000.0) );
-        assert!( close_enough_to(wrangel_east.distance_to(wrangel_west), 108000.0, 1000.0) );
+        assert!( close_enough_to(wrangel_west.distance_to(&wrangel_east), 108000.0, 1000.0) );
+        assert!( close_enough_to(wrangel_east.distance_to(&wrangel_west), 108000.0, 1000.0) );
 
-        assert!( close_enough_to(wrangel_west.bearing_to(wrangel_east), 90.0, 2.0) );
-        assert!( close_enough_to(wrangel_east.bearing_to(wrangel_west), 270.0, 2.0) );
+        assert!( close_enough_to(wrangel_west.bearing_to(&wrangel_east), 90.0, 2.0) );
+        assert!( close_enough_to(wrangel_east.bearing_to(&wrangel_west), 270.0, 2.0) );
 
         // Distance tests based on simple Google searches, like "distance from visby to melbourne"
-        assert!( close_enough_to(visby.distance_to(melbourne), 15601000.0, 10000.0) );
-        assert!( close_enough_to(melbourne.distance_to(ushuaia), 9234000.0, 10000.0) );
-        assert!( close_enough_to(ushuaia.distance_to(melbourne), melbourne.distance_to(ushuaia), 1.0) );
+        assert!( close_enough_to(visby.distance_to(&melbourne), 15601000.0, 10000.0) );
+        assert!( close_enough_to(melbourne.distance_to(&ushuaia), 9234000.0, 10000.0) );
+        assert!( close_enough_to(ushuaia.distance_to(&melbourne), melbourne.distance_to(&ushuaia), 1.0) );
+    }
+
+    #[test]
+    fn test_location_times() {
+        let utc = chrono::offset::utc::UTC::now();
+
+        //
+        // Porkkala - Naissaar
+        //
+        let porkkala_dt_fixed = chrono::DateTime::parse_from_rfc3339("2017-07-15T06:00:00+02:00").unwrap();
+        let porkkala = Location::new_with_all(59.983333, 24.433333, None, 
+            Some(porkkala_dt_fixed.with_timezone(&utc.timezone())));
+        
+        let naissaar_dt_fixed = chrono::DateTime::parse_from_rfc3339("2017-07-15T14:00:00+02:00").unwrap();
+        let naissaar = Location::new_with_all(59.566667, 24.516667, None, 
+            Some(naissaar_dt_fixed.with_timezone(&utc.timezone())));
+        
+        assert!( close_enough_to(porkkala.distance_to(&naissaar), 47000.0, 1000.0) );
+        assert!( close_enough_to(porkkala.delta_time(&naissaar).unwrap(), 8.0 * 3600.0, 0.1) );
+        assert!( close_enough_to(porkkala.average_speed(&naissaar).unwrap(), 1.63, 1.0) );
+        
+        //
+        // Naissaar - Tallin
+        //
+        let tallinn = Location::new(59.437222, 24.745278);
+        assert!( naissaar.delta_time(&tallinn).is_none() );
+        assert!( naissaar.average_speed(&tallinn).is_none() );
     }
 
     #[test]
     fn test_geoboxes() {
         let globe179 = GeoBox::new(
-            Location::from_str("90°N 179°W").unwrap(), 
-            Location::from_str("90°S 179°E").unwrap());
+            Location::new_with_str("90°N 179°W").unwrap(), 
+            Location::new_with_str("90°S 179°E").unwrap());
 
         let mediterranean = GeoBox::new(
-            Location::from_str("46°N 5°E").unwrap(), 
-            Location::from_str("30°N 37°E").unwrap());
+            Location::new_with_str("46°N 5°E").unwrap(), 
+            Location::new_with_str("30°N 37°E").unwrap());
 
         let sardinia = GeoBox::new(
-            Location::from_str("41.25°N 8°E").unwrap(), 
-            Location::from_str("39°N 10°E").unwrap());
+            Location::new_with_str("41.25°N 8°E").unwrap(), 
+            Location::new_with_str("39°N 10°E").unwrap());
             
         let gotland = GeoBox::new(
-            Location::from_str("58°N 18°E").unwrap(), 
-            Location::from_str("57°N 19.5°E").unwrap());
+            Location::new_with_str("58°N 18°E").unwrap(), 
+            Location::new_with_str("57°N 19.5°E").unwrap());
 
         let azores = GeoBox::new(
-            Location::from_str("40°N 32°W").unwrap(), 
-            Location::from_str("36°S 24°W").unwrap());
+            Location::new_with_str("40°N 32°W").unwrap(), 
+            Location::new_with_str("36°S 24°W").unwrap());
 
         let pacific_ocean = GeoBox::new(
-            Location::from_str("61°N 142°E").unwrap(), 
-            Location::from_str("75°S 69°W").unwrap());
+            Location::new_with_str("61°N 142°E").unwrap(), 
+            Location::new_with_str("75°S 69°W").unwrap());
 
         let new_zealand = GeoBox::new(
-            Location::from_str("34°S 166°E").unwrap(), 
-            Location::from_str("47°S 179°W").unwrap());
+            Location::new_with_str("34°S 166°E").unwrap(), 
+            Location::new_with_str("47°S 179°W").unwrap());
             
         let easter_island = GeoBox::new(
-            Location::from_str("27°S 110°W").unwrap(), 
-            Location::from_str("28°S 109°W").unwrap());
+            Location::new_with_str("27°S 110°W").unwrap(), 
+            Location::new_with_str("28°S 109°W").unwrap());
 
-        let taveuni = Location::from_str("16.8°S 179.5°W").unwrap();
+        let taveuni = Location::new_with_str("16.8°S 179.5°W").unwrap();
         
-        assert!( mediterranean.northwest.west_from(mediterranean.southeast) );
-        assert!( pacific_ocean.northwest.west_from(pacific_ocean.southeast) );
+        assert!( mediterranean.northwest.west_from(&mediterranean.southeast) );
+        assert!( pacific_ocean.northwest.west_from(&pacific_ocean.southeast) );
 
-        assert!( mediterranean.contains(sardinia.northwest) );
-        assert!( ! mediterranean.contains(gotland.northwest) );
+        assert!( mediterranean.contains(&sardinia.northwest) );
+        assert!( ! mediterranean.contains(&gotland.northwest) );
             
-        assert!( mediterranean.intersects(sardinia) );
-        assert!( ! mediterranean.intersects(gotland) );
-        assert!( ! mediterranean.intersects(azores) );
-        assert!( ! mediterranean.intersects(easter_island) );
-        assert!( ! mediterranean.intersects(new_zealand) );
+        assert!( mediterranean.intersects(&sardinia) );
+        assert!( ! mediterranean.intersects(&gotland) );
+        assert!( ! mediterranean.intersects(&azores) );
+        assert!( ! mediterranean.intersects(&easter_island) );
+        assert!( ! mediterranean.intersects(&new_zealand) );
         
-        assert!( pacific_ocean.contains(taveuni) );
+        assert!( pacific_ocean.contains(&taveuni) );
 
-        assert!( ! pacific_ocean.intersects(sardinia) );
-        assert!( ! pacific_ocean.intersects(gotland) );
-        assert!( ! pacific_ocean.intersects(azores) );
-        assert!( pacific_ocean.intersects(easter_island) );
-        assert!( pacific_ocean.intersects(new_zealand) );
-        assert!( ! pacific_ocean.intersects(mediterranean) );
+        assert!( ! pacific_ocean.intersects(&sardinia) );
+        assert!( ! pacific_ocean.intersects(&gotland) );
+        assert!( ! pacific_ocean.intersects(&azores) );
+        assert!( pacific_ocean.intersects(&easter_island) );
+        assert!( pacific_ocean.intersects(&new_zealand) );
+        assert!( ! pacific_ocean.intersects(&mediterranean) );
 
-        assert!( globe179.northwest.east_from(globe179.southeast) );
-        assert!( ! globe179.contains(taveuni) );
-        assert!( globe179.contains(mediterranean.northwest) );
-        assert!( globe179.intersects(mediterranean) );
-        assert!( globe179.intersects(pacific_ocean) );  
+        assert!( globe179.northwest.east_from(&globe179.southeast) );
+        assert!( ! globe179.contains(&taveuni) );
+        assert!( globe179.contains(&mediterranean.northwest) );
+        assert!( globe179.intersects(&mediterranean) );
+        assert!( globe179.intersects(&pacific_ocean) );  
     }
-
-/*
-    #[test]
-    fn test_geobox_cmp() {
-        let mut bset: BTreeSet<GeoBox> = BTreeSet::new();
-        let view = GeoBox::new(
-            Location::from_str("10°N 0°E").unwrap(),
-            Location::from_str("20°N 10°E").unwrap(),
-        );
-        bset.insert(view);
-        // TODO
-    }
-*/
 
     #[test]
     fn test_projection() {